requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via
    #   cattrs
    #   requests-cache
build==0.10.0
    # via pip-tools
cattrs==23.1.2
    # via requests-cache
certifi==2023.5.7
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   qrcode
django==4.2.3
    # via
    #   -r requirements.in
    #   django-ckeditor
    #   django-debug-toolbar
    #   django-formtools
    #   django-js-asset
    #   django-otp
    #   django-phonenumber-field
    #   django-timezone-field
    #   django-two-factor-auth
django-ckeditor==6.6.1
    # via -r requirements.in
django-debug-toolbar==4.1.0
    # via -r requirements.in
django-formtools==2.4.1
    # via django-two-factor-auth
django-js-asset==2.1.0
    # via django-ckeditor
django-otp==1.2.2
    # via django-two-factor-auth
django-phonenumber-field==6.4.0
    # via django-two-factor-auth
django-timezone-field==5.1
    # via -r requirements.in
django-two-factor-auth==1.15.3
    # via -r requirements.in
django-widget-tweaks==1.4.12
    # via -r requirements.in
exceptiongroup==1.1.2
    # via cattrs
idna==3.4
    # via requests
packaging==23.1
    # via
    #   build
    #   pip-review
pillow==10.0.0
    # via -r requirements.in
pip-review==1.3.0
    # via -r requirements.in
pip-tools==7.1.0
    # via -r requirements.in
platformdirs==3.9.1
    # via requests-cache
psycopg[binary]==3.1.9
    # via -r requirements.in
psycopg-binary==3.1.9
    # via psycopg
pypng==0.20220715.0
    # via qrcode
pyproject-hooks==1.0.0
    # via build
python-decouple==3.8
    # via -r requirements.in
pytz==2023.3
    # via django-timezone-field
qrcode==7.4.2
    # via django-two-factor-auth
requests==2.31.0
    # via
    #   -r requirements.in
    #   requests-cache
requests-cache==1.1.0
    # via -r requirements.in
six==1.16.0
    # via url-normalize
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.7.1
    # via
    #   asgiref
    #   cattrs
    #   psycopg
    #   qrcode
tzdata==2023.3
    # via
    #   django
    #   psycopg
url-normalize==1.4.3
    # via requests-cache
urllib3==2.0.4
    # via
    #   requests
    #   requests-cache
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
